Remove unneccessary boxing
No functional changes.

+++ ryhmasahkoposti-service/src/test/java/fi/vm/sade/ryhmasahkoposti/service/GroupEmailReportingServiceTest.java
@@ -64,7 +64,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.invocation.InvocationOnMock;
 import org.mockito.stubbing.Answer;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
@@ -214,7 +213,7 @@ public class GroupEmailReportingServiceTest {
         Long liiteID = groupEmailReportingService.saveAttachment(mockedFileItem);
 
         assertNotNull(liiteID);
-        assertTrue(liiteID.longValue() > 0);
+        assertTrue(liiteID > 0);
     }
 
     @Test
